+++ Assets/OkapiKit/Scripts/Editor/ActionProbabilityDrawer.cs
@@ -12,6 +12,26 @@ namespace OkapiKit.Editor
         // Draw the property inside the given rect
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
+            var propAction = property.FindPropertyRelative("action");
+
+            if (Event.current.type == EventType.ContextClick)
+            {
+                if (position.Contains(Event.current.mousePosition))
+                {
+                    // Create the context menu
+                    GenericMenu menu = new GenericMenu();
+
+                    // Add custom menu items
+                    menu.AddItem(new GUIContent("Ping"), false, () => { propAction.isExpanded = true; PingAction(propAction.objectReferenceValue as Action); });
+
+                    // Show the menu
+                    menu.ShowAsContext();
+
+                    // Use the event
+                    Event.current.Use();
+                }
+            }
+
             // Using BeginProperty / EndProperty on the parent property means that
             // prefab override logic works on the entire property.
             EditorGUI.BeginProperty(position, label, property);
@@ -27,7 +47,6 @@ namespace OkapiKit.Editor
             var actionRect = new Rect(position.x, position.y, position.width - 60, 18);
             var delayRect = new Rect(position.x + position.width - 60, position.y, 50, 18);
 
-            var propAction = property.FindPropertyRelative("action");
             EditorGUI.PropertyField(actionRect, propAction, GUIContent.none);
 
             if (propAction.objectReferenceValue != null)
@@ -48,6 +67,16 @@ namespace OkapiKit.Editor
             EditorGUI.indentLevel = indent;
 
             EditorGUI.EndProperty();
+        }
+        private void PingAction(Action action)
+        {
+            if (action == null) return;
+
+            EditorGUIUtility.PingObject(action);
+
+            // Select object
+            Selection.activeObject = action.gameObject;
+            OkapiConfig.PingComponent(action);
         }
 
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label)